Add support for GET methods.
Add support for HTTP GET method.
Mapping between the http GET and POST method is done by prefixing the
the keyword to the correct slot. For example, given a GET request /example,
the slot should be named get_example, so that proper routing can be
made.

Also, the GET slot receives as argument the query string as a
QVariantMap.